backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
websockets==15.0.1
websocket-client==1.8.0
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
